Shifting UseBlockChainData to TS
